requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
onnx==1.18.0
onnxruntime==1.22.1
onnxoptimizer==0.3.13

# FastAPI and Web Framework
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
multipart==1.3.0
pydantic==2.11.7
pydantic-settings==2.10.1
tqdm==4.67.1
# Image Processing
pillow==11.3.0
opencv-python==4.12.0.88

# Data Science and Utilities
pandas==2.3.1
scikit-learn==1.7.1
matplotlib==3.10.5
seaborn==0.13.2

# Configuration and Logging
python-dotenv==1.0.0
PyYAML==6.0.2
loguru==0.7.3

# Development and Testing
pytest==8.4.1
pytest-asyncio==1.1.0
black==25.1.0
flake8==7.3.0
mypy==1.17.1

# Production and Monitoring
gunicorn==23.0.0
prometheus-client==0.22.1